% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggserif.R
\name{ggserif}
\alias{ggserif}
\title{Personal `ggplot2` theme with serif fonts and directed axes}
\usage{
ggserif(base_size = 14)
}
\arguments{
\item{base_size}{Base size of the plot. It can be modified but I usually do not touch it.}
}
\value{
NULL The function doesn't return any value.
}
\description{
This is a ggplot() theme with serif fonts and directed arrows.
}
\examples{
# Set theme globally by
ggplot2::theme_set(ggserif())

# Or use it for a single plot
ggplot2::ggplot(iris, ggplot2::aes(x = Sepal.Length, y = Sepal.Width, colour = Species)) +
ggplot2::geom_point() +
ggplot2::theme_set(ggserif())
}
