% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rm_feature.R
\name{na_rm_feature}
\alias{na_rm_feature}
\title{Removing Rows Based on Missing Values in a Column}
\usage{
na_rm_feature(df, colname)
}
\arguments{
\item{df}{Data frame to remove observations from}

\item{colname}{Column name to check for missing values}
}
\value{
Data frame with no missing values in column colname
}
\description{
Sometimes, I do not want to na.omit() because it will treat all features equally. I
want to check values only for one column, while removing those observations.
}
