% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_zero.R
\name{na_to_zero}
\alias{na_to_zero}
\title{Convert missing values to zero}
\usage{
na_to_zero(x)
}
\arguments{
\item{x}{Vector}
}
\value{
y Manipulated vector
}
\description{
This function converts missing values in a vector to zero.
}
\examples{
iris$Sepal.Width[2] = NA
na_to_zero(iris$Sepal.Width)
}
