% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_na.R
\name{which_na}
\alias{which_na}
\title{Which values are missing?}
\usage{
which_na(x)
}
\arguments{
\item{x}{Vector to find missing elements}
}
\value{
Index of elements missing
}
\description{
This function returns index of elements which are missing.
}
\examples{
l = c(0, 1, NA, 3, 4, 5)
which_na(l)
}
